backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3==1.39.3
requests-oauthlib==2.0.0
cryptography==45.0.6
python-dotenv==1.1.1
pymongo==4.7.0
pydantic==2.9.2
email-validator==2.2.0
pyjwt==2.10.1
passlib==1.7.4
tzdata==2025.2
motor==3.3.1
pytest==8.4.1
black==25.1.0
isort==6.0.1
flake8==7.3.0
mypy==1.17.1

requests==2.32.4
httpx==0.28.1
pandas==2.3.2
numpy==2.3.1
python-multipart==0.0.20
jq==1.10.0
typer==0.16.0
mongomock==4.3.0
safety==3.6.0
